Snake AI using DQN and basic information(low performance)

+++ BaseNet.py
@@ -6,10 +6,10 @@ import numpy
 class BaseNet(nn.Module):
     def __init__(self):
         super(BaseNet, self).__init__()
-        self.fc1 = nn.Linear(6, 512)
-        self.fc2 = nn.Linear(512, 256)
-        self.fc3 = nn.Linear(256, 128)
-        self.fc4 = nn.Linear(128, 1)
+        self.fc1 = nn.Linear(6, 512, dtype=torch.float64)
+        self.fc2 = nn.Linear(512, 256, dtype=torch.float64)
+        self.fc3 = nn.Linear(256, 128, dtype=torch.float64)
+        self.fc4 = nn.Linear(128, 1, dtype=torch.float64)
     
     def forward(self, x):
         x = self.fc1(x)
